Clean up include tree
