+++ dGame/dComponents/ScriptedActivityComponent.cpp
@@ -18,6 +18,7 @@
 #include "dConfig.h"
 #include "InventoryComponent.h"
 #include "DestroyableComponent.h"
+#include "dMessageIdentifiers.h"
 
 ScriptedActivityComponent::ScriptedActivityComponent(Entity* parent, int activityID) : Component(parent) {
 	m_ActivityID = activityID;
