+++ dScripts/02_server/Map/General/PropertyPlatform.cpp
@@ -1,6 +1,7 @@
 #include "PropertyPlatform.h"
 #include "RebuildComponent.h"
 #include "GameMessages.h"
+#include "MovingPlatformComponent.h"
 
 void PropertyPlatform::OnRebuildComplete(Entity* self, Entity* target) {
 	//    auto* movingPlatform = self->GetComponent<MovingPlatformComponent>();
@@ -9,7 +10,7 @@ void PropertyPlatform::OnRebuildComplete(Entity* self, Entity* target) {
 	//        movingPlatform->SetNoAutoStart(true);
 	//    }
 	GameMessages::SendPlatformResync(self, UNASSIGNED_SYSTEM_ADDRESS, true, 0,
-		0, 0, MovementPlatformState::Stationary);
+		0, 0, eMovementPlatformState::Stationary);
 }
 
 void PropertyPlatform::OnUse(Entity* self, Entity* user) {
@@ -20,7 +21,7 @@ void PropertyPlatform::OnUse(Entity* self, Entity* user) {
 		//            movingPlatform->GotoWaypoint(1);
 		//        }
 		GameMessages::SendPlatformResync(self, UNASSIGNED_SYSTEM_ADDRESS, true, 0,
-			1, 1, MovementPlatformState::Moving);
+			1, 1, eMovementPlatformState::Moving);
 
 		self->AddCallbackTimer(movementDelay + effectDelay, [self, this]() {
 			self->SetNetworkVar<float_t>(u"startEffect", dieDelay);
